Deprecate RequestListCaster
